requirements.txt
# ==============================================================================
# Sistema Empresarial E-commerce - Clean Architecture
# Python 3.14+ | Deploy: Hostinger VPS
# ==============================================================================

# ------------------------------------------------------------------------------
# Web Frameworks
# ------------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
django>=5.1.0
djangorestframework>=3.15.0

# ------------------------------------------------------------------------------
# Database - PostgreSQL (PRODUCTION)
# ------------------------------------------------------------------------------
psycopg[binary]>=3.2.0        # PostgreSQL adapter (psycopg3 - Recomendado)
# psycopg2-binary>=2.9.9      # Alternativa: psycopg2 (legacy, solo si es necesario)
python-dotenv>=1.0.0          # Cargar variables de entorno desde .env

# ------------------------------------------------------------------------------
# Database - MySQL (Legacy/Desarrollo)
# ------------------------------------------------------------------------------
# pymysql>=1.1.0              # MySQL adapter (comentado, usar PostgreSQL)
# cryptography>=43.0.0        # Requerido para PyMySQL con SSL

# ------------------------------------------------------------------------------
# ORM y Database Toolkit
# ------------------------------------------------------------------------------
sqlalchemy>=2.0.0             # ORM independiente (alternativa a Django ORM)

# ------------------------------------------------------------------------------
# Validación y Serialización
# ------------------------------------------------------------------------------
pydantic>=2.9.0
pydantic-settings>=2.6.0      # Gestión de configuración
email-validator>=2.2.0        # Validación de emails

# ------------------------------------------------------------------------------
# Seguridad
# ------------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0    # JWT tokens
passlib[bcrypt]>=1.7.4              # Hashing de passwords
python-multipart>=0.0.12            # Form data (uploads)

# ------------------------------------------------------------------------------
# Utilidades
# ------------------------------------------------------------------------------
python-dotenv>=1.0.1          # Variables de entorno
httpx>=0.27.0                 # Cliente HTTP async
tenacity>=9.0.0               # Retry logic

# ------------------------------------------------------------------------------
# Logging y Monitoreo
# ------------------------------------------------------------------------------
structlog>=24.4.0             # Logging estructurado
rich>=13.9.0                  # Console output mejorado

# ------------------------------------------------------------------------------
# ASGI/WSGI Server (Producción Hostinger)
# ------------------------------------------------------------------------------
gunicorn>=23.0.0              # WSGI server para Django
hypercorn>=0.17.0             # ASGI server alternativo

# ------------------------------------------------------------------------------
# Development (comentar en producción)
# ------------------------------------------------------------------------------
# pytest>=8.3.0
# pytest-asyncio>=0.24.0
# pytest-cov>=6.0.0
# black>=24.10.0
# ruff>=0.7.0
# mypy>=1.13.0
# pre-commit>=4.0.0
